64-bit changes
SCCS-vsn: 8.3

+++ sys/netiso/iso.c
@@ -30,7 +30,7 @@
  * OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF
  * SUCH DAMAGE.
  *
- *	@(#)iso.c	8.2 (Berkeley) 11/15/93
+ *	@(#)iso.c	8.3 (Berkeley) 01/09/95
  */
 
 /***********************************************************
@@ -413,7 +413,7 @@ caddr_t			buf;		/* RESULT: network portion of address here */
 /* ARGSUSED */
 iso_control(so, cmd, data, ifp)
 	struct socket *so;
-	int cmd;
+	u_long cmd;
 	caddr_t data;
 	register struct ifnet *ifp;
 {
